+++ src/main/java/com/ururulab/ururu/groupBuy/controller/dto/request/GroupBuyStatusUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.ururulab.ururu.groupBuy.domain.dto.request;
+package com.ururulab.ururu.groupBuy.controller.dto.request;
 
 import com.ururulab.ururu.groupBuy.domain.entity.enumerated.GroupBuyStatus;
 import jakarta.validation.constraints.NotNull;
