+++ src/main/java/com/ururulab/ururu/groupBuy/domain/entity/GroupBuy.java
@@ -12,11 +12,8 @@ import org.hibernate.annotations.JdbcTypeCode;
 import org.hibernate.type.SqlTypes;
 
 import java.time.Instant;
-import java.time.ZonedDateTime;
 import java.util.ArrayList;
 import java.util.List;
-
-import static com.ururulab.ururu.groupBuy.domain.dto.validation.GroupBuyValidationConstants.*;
 
 @Entity
 @Getter
