+++ src/main/java/com/ururulab/ururu/groupBuy/service/validation/GroupBuyValidator.java
@@ -1,8 +1,8 @@
 package com.ururulab.ururu.groupBuy.service.validation;
 
 import com.ururulab.ururu.global.exception.BusinessException;
-import com.ururulab.ururu.groupBuy.domain.dto.request.GroupBuyOptionRequest;
-import com.ururulab.ururu.groupBuy.domain.dto.request.GroupBuyRequest;
+import com.ururulab.ururu.groupBuy.controller.dto.request.GroupBuyOptionRequest;
+import com.ururulab.ururu.groupBuy.controller.dto.request.GroupBuyRequest;
 import com.ururulab.ururu.groupBuy.domain.repository.GroupBuyRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
