+++ src/main/java/com/ururulab/ururu/product/controller/ProductController.java
@@ -1,21 +1,16 @@
 package com.ururulab.ururu.product.controller;
 
-import com.nimbusds.jwt.JWT;
 import com.ururulab.ururu.global.domain.dto.ApiResponseFormat;
-import com.ururulab.ururu.product.domain.dto.request.ProductRequest;
-import com.ururulab.ururu.product.domain.dto.response.ProductListResponse;
-import com.ururulab.ururu.product.domain.dto.response.ProductOptionResponse;
-import com.ururulab.ururu.product.domain.dto.response.ProductResponse;
+import com.ururulab.ururu.product.controller.dto.request.ProductRequest;
+import com.ururulab.ururu.product.controller.dto.response.ProductListResponse;
+import com.ururulab.ururu.product.controller.dto.response.ProductOptionResponse;
+import com.ururulab.ururu.product.controller.dto.response.ProductResponse;
 import com.ururulab.ururu.product.service.ProductOptionService;
 import com.ururulab.ururu.product.service.ProductService;
 import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.media.Content;
-import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
