+++ src/main/java/com/ururulab/ururu/product/controller/dto/response/ProductOptionResponse.java
@@ -1,4 +1,4 @@
-package com.ururulab.ururu.product.domain.dto.response;
+package com.ururulab.ururu.product.controller.dto.response;
 
 import com.ururulab.ururu.product.domain.entity.ProductOption;
