+++ src/main/java/com/ururulab/ururu/product/domain/entity/ProductNotice.java
@@ -6,8 +6,6 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import static com.ururulab.ururu.product.domain.dto.validation.ProductValidationConstants.*;
-
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
