+++ src/main/java/com/ururulab/ururu/product/service/ProductOptionService.java
@@ -2,9 +2,9 @@ package com.ururulab.ururu.product.service;
 
 import com.ururulab.ururu.global.exception.BusinessException;
 import com.ururulab.ururu.global.exception.error.ErrorCode;
-import com.ururulab.ururu.product.domain.dto.request.ProductImageUploadRequest;
-import com.ururulab.ururu.product.domain.dto.request.ProductOptionRequest;
-import com.ururulab.ururu.product.domain.dto.response.ProductOptionResponse;
+import com.ururulab.ururu.product.controller.dto.request.ProductImageUploadRequest;
+import com.ururulab.ururu.product.controller.dto.request.ProductOptionRequest;
+import com.ururulab.ururu.product.controller.dto.response.ProductOptionResponse;
 import com.ururulab.ururu.product.domain.entity.Product;
 import com.ururulab.ururu.product.domain.entity.ProductOption;
 import com.ururulab.ururu.product.domain.entity.enumerated.Status;
