+++ src/main/java/com/ururulab/ururu/product/service/ProductService.java
@@ -3,10 +3,10 @@ package com.ururulab.ururu.product.service;
 import com.ururulab.ururu.global.domain.entity.TagCategory;
 import com.ururulab.ururu.global.exception.BusinessException;
 import com.ururulab.ururu.global.exception.error.ErrorCode;
-import com.ururulab.ururu.product.domain.dto.request.ProductImageUploadRequest;
-import com.ururulab.ururu.product.domain.dto.request.ProductNoticeRequest;
-import com.ururulab.ururu.product.domain.dto.request.ProductRequest;
-import com.ururulab.ururu.product.domain.dto.response.*;
+import com.ururulab.ururu.product.controller.dto.request.ProductImageUploadRequest;
+import com.ururulab.ururu.product.controller.dto.request.ProductNoticeRequest;
+import com.ururulab.ururu.product.controller.dto.request.ProductRequest;
+import com.ururulab.ururu.product.controller.dto.response.*;
 import com.ururulab.ururu.product.domain.entity.*;
 import com.ururulab.ururu.product.domain.entity.enumerated.Status;
 import com.ururulab.ururu.product.domain.repository.*;
